gg
change login

+++ app/src/main/java/com/tiha/anphat/ui/cart/CartContract.java
@@ -27,7 +27,7 @@ public interface CartContract {
 
         void onGetDonGiaProductByUserError(String error);
 
-        void onGetProductInventorySuccess(Integer result);
+        void onGetProductInventorySuccess(Double result);
 
         void onGetProductInventoryError(String error);
