+++ app/src/main/java/com/tiha/anphat/ui/product/ProductPresenter.java
@@ -71,7 +71,7 @@ public class ProductPresenter implements ProductContract.Presenter {
     public void GetProductInventory(String maKho, String productID, String date) {
         model.GetProductInventory(maKho, productID, date, new IProductModel.IGetProductInventoryFinish() {
             @Override
-            public void onSuccess(Integer result) {
+            public void onSuccess(Double result) {
                 view.onGetProductInventorySuccess(result);
             }
